Added project main jar file, data & library
This commit adds source code jar file  & a new data folder to the project containing food data, as well as the Jackson library in a JAR file. The Jackson library is used for parsing and serializing JSON data. This addition will enable the application to read and write data in a structured format and improve its functionality.